move VirtualService out of resource pkg
