+++ internal/istio/virtualservice.go
@@ -1,4 +1,4 @@
-package resource
+package istio
 
 import (
 	"errors"
